+++ src/components/author/AuthorItems.jsx
@@ -1,19 +1,36 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import AuthorImage from "../../images/author_thumbnail.jpg";
-import nftImage from "../../images/nftImage.jpg";
+import React, { useState, useEffect } from "react";
+import { Link, useParams } from "react-router-dom";
+import axios from "axios";
 
-const AuthorItems = () => {
+const AuthorItems = ({ nftData }) => {
+  const [localNftData, setLocalNftData] = useState([]);
+  const { id } = useParams();
+
+  useEffect(() => {
+    const fetchNFTs = async () => {
+      try {
+        const response = await axios.get(
+          `https://us-central1-nft-cloud-functions.cloudfunctions.net/authors/${id}` 
+        );
+        setLocalNftData(response.data);
+      } catch (error) {
+        console.error("There was an error fetching the NFT data!", error);
+      }
+    };
+    fetchNFTs();
+  }, [id]);
+
   return (
     <div className="de_tab_content">
       <div className="tab-1">
         <div className="row">
-          {new Array(8).fill(0).map((_, index) => (
+          
+        {localNftData.map((nft, index) => (
             <div className="col-lg-3 col-md-6 col-sm-6 col-xs-12" key={index}>
               <div className="nft__item">
                 <div className="author_list_pp">
                   <Link to="">
-                    <img className="lazy" src={AuthorImage} alt="" />
+                    <img className="lazy" src={nft.authorImage} alt="" />
                     <i className="fa fa-check"></i>
                   </Link>
                 </div>
@@ -35,22 +52,22 @@ const AuthorItems = () => {
                       </div>
                     </div>
                   </div>
-                  <Link to="/item-details">
+                  <Link to={`/item-details/:${id}`}>
                     <img
-                      src={nftImage}
+                      src={nft.nftImage}
                       className="lazy nft__item_preview"
                       alt=""
                     />
                   </Link>
                 </div>
                 <div className="nft__item_info">
-                  <Link to="/item-details">
+                  <Link to={`/item-details/:${id}`}>
                     <h4>Pinky Ocean</h4>
                   </Link>
-                  <div className="nft__item_price">2.52 ETH</div>
+                  <div className="nft__item_price">{`${nft.price}`} ETH</div>
                   <div className="nft__item_like">
                     <i className="fa fa-heart"></i>
-                    <span>97</span>
+                    <span>{nft.likes}</span>
                   </div>
                 </div>
               </div>
